Fix setting the miners and values by timestamp Removed hardcoded mineers and values by timestamp Used a for loop to resolve the issue Verified there was a dynamic amount of set miners and values by timestamp

+++ contracts/zap-miner/libraries/ZapLibrary.sol
@@ -182,7 +182,7 @@ library ZapLibrary {
         emit NewValue(
             _requestId,
             self.uintVars[keccak256('timeOfLastNewValue')],
-            self.currentMiners[a[a.length/2]].value,
+            self.currentMiners[a[a.length / 2]].value,
             self.uintVars[keccak256('currentTotalTips')] -
                 (self.uintVars[keccak256('currentTotalTips')] % 5),
             self.currentChallenge
@@ -201,16 +201,22 @@ library ZapLibrary {
         //Save the official(finalValue), timestamp of it, 5 miners and their submitted values for it, and its block number
         _request.finalValues[
             self.uintVars[keccak256('timeOfLastNewValue')]
-        ] = self.currentMiners[a[a.length/2]].value;
+        ] = self.currentMiners[a[a.length / 2]].value;
         _request.requestTimestamps.push(
             self.uintVars[keccak256('timeOfLastNewValue')]
         );
 
-
         //set miners and values by timestamp
-        for(uint i = 0; i < a.length; i++) {
-            _request.minersByValue[self.uintVars[keccak256('timeOfLastNewValue')]].push(self.currentMiners[a[i]].miner);
-            _request.valuesByTimestamp[self.uintVars[keccak256('timeOfLastNewValue')]].push(self.currentMiners[a[i]].value);
+        for (uint256 i = 0; i < a.length; i++) {
+            _request.minersByValue[
+                self.uintVars[keccak256('timeOfLastNewValue')]
+            ]
+                .push(self.currentMiners[a[i]].miner);
+
+            _request.valuesByTimestamp[
+                self.uintVars[keccak256('timeOfLastNewValue')]
+            ]
+                .push(self.currentMiners[a[i]].value);
         }
 
         _request.minedBlockNum[
